Update README.md
Catch me if you can (or CatchMe) - is a supporting service for connecting students of one or more Universities. General idea of this website is to amuse the students, however, it can be a useful tool for traveling purposes (because it can content the useful and important information about some locations). At this website the user enters a hashtag in the input field and at the map he/she can see all the locations with such a hashtag. 
The peculiarity of this process is in the fact that one hashtag can lead to different location. For example, if user insert #Arctic, their could see many different places on map with the same hashtag, like photos with real Arctic places as well as pubs with name “Arctic” in South Africa or hotel “Arctic” in Murmansk or night club “Arctic” in Saint-Petersburg. 
Or the student can enter #ITMO and find where other students of ITMO have been in Saint-Petersburg or around the world . The same for the students of all universities!
The optional part of the website is to add the tracking function: the traveler can create and show to the others his/her route using the hashtags.
Problem: Students always want to be in close connection with others and want to know about interesting places as well as they want to create their own traveling plans they implement after session! (or doing the session or after exam))) 
Solution: Create and develop an webservice which will help to all students stay in touch and enjoy the communications in students community. Moreover, each other community can use this website to join and follow the #hashtags. 
Framework: Django/Flask
Simplified cases:
•	A student wants to go to the pub after an exam near the University. She/he enter #pub and could see the different places with such a hashtag. It could be real pubs, but it also could a concert of the rock band with the name “pub” and finally the student can ask others is it interesting or not and find the location of the concert (and visit it). 
•	An ITMO student is traveling around the world and add the place on the map with #hashtag “Helsinki”, thus, he can see all the places with this hashtag around the world and in Helsinki and can enjoy the sense of brotherhood or even find some other ITMO students in Helsinki for pub crawling.  
References:
1. Instagram platform Policy: https://www.instagram.com/about/legal/terms/api/
2. Instagram Developer Documentation: https://developers.facebook.com/docs/instagram-api